:art: 페이지 전체 사용 데이터 CategoryPage에서 데이터 호출 후 뿌려주는 방식으로 변경
