+++ client/src/pages/Category/index.tsx
@@ -1,5 +1,10 @@
-import React from 'react';
+import React, { useEffect } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { useParams } from 'react-router-dom';
 import { Body, SideNav } from 'src/components/Category';
+import { useRootCategoryState } from 'src/hooks';
+import { RootState } from 'src/states';
+import { getServicesByCategoryIdx } from 'src/states/service/action';
 import styled from 'styled-components';
 
 const Container = styled.div`
@@ -12,10 +17,25 @@ const Container = styled.div`
 `;
 
 const CategoryPage = () => {
+  const { categoryIdx } = useParams();
+
+  const [rootCategory, setRootCategory] = useRootCategoryState(
+    Number(categoryIdx)
+  );
+  const serviceList = useSelector(
+    (state: RootState) => state.serviceCardList.serviceCardList
+  );
+
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(getServicesByCategoryIdx(Number(categoryIdx)));
+  }, [categoryIdx, dispatch]);
+
   return (
     <Container>
-      <SideNav />
-      <Body />
+      <SideNav rootCategory={rootCategory} />
+      <Body serviceList={serviceList} />
     </Container>
   );
 };
